+++ balancer/container.sh
@@ -5,31 +5,31 @@ set -eo pipefail
 cd $(dirname $0)
 . ../common/docker_helper.sh
 
-# Generate dummy config/routers if needed
-mkdir -p config
-if [ ! -f config/routers ]; then
-  echo "127.0.0.1" > config/routers
-fi
 
 STOP_SIGNAL=SIGQUIT
 STOP_TIMEOUT=30
+BIND_IF=${BIND_IF:-"*"}
+ROUTERS_FILE=${ROUTERS_FILE:-"routers"}
 
-REGISTRY_URL=${REGISTRY_URL:-""}
-
-if [ "$REGISTRY_URL" != "" ]; then
-  REGISTRY_URL="$REGISTRY_URL""/"
+# Generate dummy config/routers if needed
+mkdir -p config
+if [ ! -f config/$ROUTERS_FILE ]; then
+  echo "127.0.0.1" > config/$ROUTERS_FILE
 fi
 
-if [ "$AIR_ROUTERS" != "" ]; then
-  DOCKER_START_ARGS="-e AIR_ROUTERS=$AIR_ROUTERS"
+DOCKER_IMAGE=$(aircloak_image_name air_balancer)
+DOCKER_START_ARGS="
+  --net=host
+  -v $(pwd)/config:/aircloak/balancer/config
+  -e BIND_IF=$BIND_IF
+  -e ROUTERS_FILE=$ROUTERS_FILE
+"
+
+if [ "$BIND_IF" != "*" ]; then
+  CONTAINER_NAME="air_balancer_$BIND_IF"
+else
+  CONTAINER_NAME="air_balancer"
 fi
+CONTAINER_ARGS="/aircloak/balancer/start.sh"
 
-DOCKER_START_ARGS="$DOCKER_START_ARGS \
-  --net=host \
-  -v $(pwd)/config:/aircloak/balancer/config \
-  "$REGISTRY_URL"aircloak/air_balancer:latest \
-  /aircloak/balancer/start.sh"
-
-REMOTE_CONSOLE_COMMAND="/bin/bash"
-
-container_ctl air_balancer $@
+container_ctl $@