+++ common/docker_helper.sh
@@ -2,17 +2,6 @@
 
 . $(dirname ${BASH_SOURCE[0]})/../config/config.sh
 
-function push_to_registry {
-  if named_container_running air_docker_registry; then
-    REGISTRY_URL="127.0.0.1:$(get_tcp_port prod registry/http)"
-    echo "Pushing $1 to local registry"
-    docker tag -f aircloak/$1:latest $REGISTRY_URL/aircloak/$1:latest
-    docker push $REGISTRY_URL/aircloak/$1:latest
-  else
-    echo "Warning: local registry is not running, image not pushed."
-  fi
-}
-
 function named_container_running {
   if [ -z "$(docker ps --filter=name=$1 | grep -v CONTAINER)" ]; then
     return 1
@@ -54,62 +43,48 @@ function stop_named_container {
 }
 
 function container_ctl {
-  container_name=$1
-  shift || true
-
   command=$1
   shift || true
 
-  if [ "$AIR_ENV" = "prod" ]; then
-    driver_arg="--log-driver=syslog --log-opt syslog-tag=$container_name"
-    container_env="-e AIR_ENV=prod"
-  fi
-
   case "$command" in
     start)
-      stop_named_container $container_name
-      echo "Starting container $container_name"
-      docker run -d $driver_arg $container_env --restart on-failure --name $container_name $DOCKER_START_ARGS
-      ;;
-
-    ensure_started)
-      if ! named_container_running $container_name ; then
-        container_ctl $container_name start "$@"
-      fi
-      ;;
-
-    ensure_latest_version_started)
-      if latest_version_running $container_name; then
-        echo "$container_name already on the latest image version"
-      else
-        echo "Restarting $container_name to ensure that the latest version is running."
-        container_ctl $container_name start "$@"
-      fi
+      DOCKER_START_ARGS="$DOCKER_START_ARGS -d $driver_arg --restart on-failure"
+      start_container
       ;;
 
     console)
-      stop_named_container $container_name
-      docker run --rm -it $container_env --name $container_name $DOCKER_START_ARGS
+      DOCKER_START_ARGS="--rm -it $DOCKER_START_ARGS"
+      start_container
       ;;
 
     foreground)
-      stop_named_container $container_name
-      docker run --rm -i $container_env --name $container_name $DOCKER_START_ARGS
+      DOCKER_START_ARGS="--rm -i $DOCKER_START_ARGS"
+      start_container
+      ;;
+
+    ensure_started)
+      if ! named_container_running $CONTAINER_NAME ; then
+        container_ctl start $@
+      fi
       ;;
 
     ssh)
-      docker exec -i -t $container_name \
-        /bin/bash -c "TERM=xterm CONTAINER_NAME='$container_name' /bin/bash -l"
+      docker exec -i -t $CONTAINER_NAME \
+        /bin/bash -c "TERM=xterm CONTAINER_NAME='$CONTAINER_NAME' /bin/bash -l"
       ;;
 
     remote_console)
-      docker exec -i -t $container_name \
-        /bin/bash -c "TERM=xterm $REMOTE_CONSOLE_COMMAND"
+      docker exec -i -t $CONTAINER_NAME \
+        /bin/bash -c "TERM=xterm ${REMOTE_CONSOLE_COMMAND:-"/bin/bash"}"
       ;;
 
     stop)
-      stop_named_container $container_name
+      stop_named_container $CONTAINER_NAME
       exit 0
+      ;;
+
+    image_name)
+      echo "$DOCKER_IMAGE"
       ;;
 
     *)
@@ -122,6 +97,7 @@ function container_ctl {
         remote_console - opens an application session (e.g. Erlang or Rails console) in the running container
         console - starts the container in foreground (interactive)
         foreground - starts the container in foreground (non-interactive)
+        image_name - prints the image name
         $CUSTOM_COMMANDS
       "
 
@@ -138,12 +114,16 @@ function container_ctl {
   esac
 }
 
-# Just like build_aircloak_image, but also versions the image and pushes it to the
-# registry.
-function build_production_image {
-  build_aircloak_image "$@"
-  push_to_registry $1
-  version_latest_image $1
+function start_container {
+  if [ "$REGISTRY_URL" != "" ]; then DOCKER_IMAGE="$REGISTRY_URL/$DOCKER_IMAGE"; fi
+
+  if [ "$DOCKER_IMAGE_VERSION" != "" ]; then
+    DOCKER_IMAGE="$DOCKER_IMAGE:$DOCKER_IMAGE_VERSION"
+  fi
+
+  stop_named_container $CONTAINER_NAME
+  echo "Starting container $CONTAINER_NAME from $DOCKER_IMAGE"
+  docker run $DOCKER_START_ARGS --name $CONTAINER_NAME $DOCKER_IMAGE $CONTAINER_ARGS
 }
 
 # Syntax:
@@ -166,12 +146,14 @@ function build_aircloak_image {
     if [ -f .dockerignore ]; then rm .dockerignore; fi
   fi
 
+  full_image_name=$(aircloak_image_name $1)
+
   temp_docker_file="tmp/$(uuidgen).dockerfile"
   {
     mkdir -p tmp
-    echo "[aircloak] building aircloak/$1"
+    echo "[aircloak] building $full_image_name"
     cat $dockerfile | dockerfile_content > "$temp_docker_file"
-    docker build -t aircloak/$1:latest -f "$temp_docker_file" .
+    docker build -t $full_image_name:latest -f "$temp_docker_file" .
   } || {
     # called in the case of an error
     exit_code=$?
@@ -183,7 +165,30 @@ function build_aircloak_image {
 
   if [ -f "$temp_docker_file" ]; then rm "$temp_docker_file"; fi
   if [ -f .dockerignore ]; then rm .dockerignore; fi
+
+  # remove exited containers
+  exited_containers=$(docker ps --quiet -f status=exited)
+  if [ "$exited_containers" != "" ]; then docker rm $exited_containers || true; fi
+
+  # remove local version tags (obsolete in the new version of the build system)
+  local_version_tags=$(docker images | awk "{if (\$1 == \"$full_image_name\" && \$2 != \"latest\") print \$1\":\"\$2}")
+  if [ "$local_version_tags" != "" ]; then docker rmi $local_version_tags || true; fi
+
+  # remove dangling images
+  dangling_images=$(docker images --quiet --filter "dangling=true")
+  if [ "$dangling_images" != "" ]; then docker rmi $dangling_images || true; fi
+
   cd $curdir
+}
+
+function aircloak_image_name {
+  if [ "$IMAGE_CATEGORY" == "" ]; then
+    image_name="$1"
+  else
+    image_name="${IMAGE_CATEGORY}_$1"
+  fi
+
+  echo "aircloak/$image_name"
 }
 
 # Produces the final dockerfile contents.
@@ -225,7 +230,7 @@ function air_init {
   # Support for rebuilding of all images. Change this date if you want to
   # force the rebuild of all images. Usually you want to do this if you
   # want to upgrade Debian packages on all images.
-  upgrade_date="20160218"
+  upgrade_date="20160302"
   echo "RUN echo '$upgrade_date' > /dev/null"
 
   # Start the RUN command
@@ -248,18 +253,6 @@ EOF
   # Upgrade existing packages
   echo "apt-get update && apt-get upgrade -y && \\ "
 
-  # Generate special useradd helper which ensures proper UID on Linux hosts
-  if [ "$(which docker-machine)" == "" ]; then
-    # No docker-machine -> use host's $UID to ensure proper permissions
-    uid_arg="-u $UID"
-  fi
-
-  cat <<-EOF
-    echo '#!/usr/bin/env bash' > /tmp/build_config/useradd.sh && \\
-    echo 'useradd $uid_arg "\$@"' >> /tmp/build_config/useradd.sh && \\
-    chmod +x /tmp/build_config/useradd.sh && \\
-EOF
-
   # Support for setting custom prompt through CONTAINER_NAME env. This allows us
   # to inject the container name into the ssh session. Without this, the name of the
   # real host machine will be used, since containers run in the shared networking mode.
@@ -267,50 +260,6 @@ EOF
     echo 'PS1="\${debian_chroot:+(\$debian_chroot)}\\\\u@\${CONTAINER_NAME:-\\\\h}:\\\\w# "' \\
       >> /etc/profile.d/custom_prompt.sh
 EOF
-}
-
-function version_latest_image {
-  image_name="aircloak/$1"
-  latest_image_id=$(find_images $image_name ^latest$)
-  if [ "$latest_image_id" == "" ]; then
-    echo "Can't find latest image for $image_name"
-    return 1
-  fi
-
-  major_minor=$(cat $(dirname ${BASH_SOURCE[0]})/../VERSION)
-
-  last_patch_version=$(
-        find_images $image_name ^$major_minor.[0-9]+$ version |
-            sed "s/$major_minor\.//" |
-            sort -r -n |
-            head -n 1
-      )
-  latest_version_image_id=$(find_images $image_name ^$major_minor.$last_patch_version$)
-
-  if [ "$latest_version_image_id" == "$latest_image_id" ]; then
-    echo "No changes for $image_name"
-    return 0
-  fi
-
-  # Tag with major.minor.last_patch+1
-  # If there's no patch version for current major.minor (as defined in VERSION file),
-  # we'll tag with major.minor.0
-  patch_version=$((${last_patch_version:-"-1"} + 1))
-  tag="$image_name:$major_minor.$patch_version"
-  echo "Tagging $latest_image_id with $tag"
-  docker tag $latest_image_id $tag
-
-  # Keep only last 5 versions.
-  for old_version in $(
-      find_images $image_name "^[0-9]+\.[0-9]+\.[0-9]+$" version |
-          # sort by version descending (using numeric comparison for each field)
-          sort -t "." -k "1,1rn" -k "2,2rn" -k "3,3rn" |
-          # skip 5 most recent versions
-          tail -n+6
-      )
-  do
-    docker rmi "$image_name:$old_version"
-  done
 }
 
 function find_images {
@@ -329,30 +278,6 @@ function find_images {
       awk "{if (\$1 == \"$1\" && \$2 ~ /${2:-".*"}/) print $print_column}" | \
       sort | \
       uniq
-}
-
-function print_most_recent_versions {
-  all_images=$(docker images --no-trunc)
-  aircloak_images=$(
-        echo "$all_images" |
-            awk '{if ($1 ~ /^aircloak\/.+/) print $1}' |
-            sort |
-            uniq
-      )
-  printf "\nLatest images versions:\n"
-  for aircloak_image in $aircloak_images; do
-    latest_version=$(
-        ALL_IMAGES="$all_images" find_images $aircloak_image '^[0-9]+\.[0-9]+\.[0-9]+$' version |
-            sort -t "." -k "1,1rn" -k "2,2rn" -k "3,3rn" |
-            head -n 1
-      )
-
-      if [ "$latest_version" != "" ]; then
-        echo "$all_images" | \
-            awk "{if (\$1 == \"$aircloak_image\" && \$2 == \"$latest_version\") {\$3=\"\"; print;}}"
-      fi
-  done
-  printf "\n"
 }
 
 function latest_version_running {