+++ config/deploy/stage.rb
@@ -1,3 +1,0 @@
-role :app, %w{deployer@acdocker1.mpi-sws.org}
-role :app_admin, %w{root@acdocker1.mpi-sws.org}
-role :build, %w{deployer@acdocker1.mpi-sws.org}